requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-mysql==3.11.1
mysqlclient==2.0.3
django-cors-headers==3.7.0
PyJWT==2.1.0
python-decouple==3.4
Pillow==8.2.0
django-filter==2.4.0
psycopg2==2.8.6
django-redis==5.0.0